yet another Guidebook update
A couple of \e to \\ changes, a whole bunch of ^ to \(ha changes
(similar character but bigger so easier to see), ~ to \(ti (ditto),
and - to \- (first is a hyphen, second is a minus sign which is
bigger; see the difference for "- and | The walls of a room...").

As well as horizontal walls, open doors, top and bottom of swallow
and explode octagons, I changed command line switches -s, -X, and -D
to use minus sign.  I was unsure whether the umpteen M-C should be
changed too.  Since it was less work to leave them as-is, that's
what I did.  I also left ^C with smaller circumflex punctation
instead of changing to bigger circumflex character for same reason.

The [much smaller] Guidebook.tex changes need testing.

Not fixed:
I think the sentence "you can type 'nethack -s all' on most versions"
is very misleading.  Using 'versions' to refer to the various ports
rather than different releases is iffy, but my complaint is that
"typing something" sounds like an action you would perform while in
the game.  Access to a command line and figuring out how to invoke
nethack from there is probably not something where "most versions"
applies any more.  But I don't know how to rephrase that succinctly.
If I did, I would have just changed it....